requirements.txt
streamlit
extra_streamlit_components
numpy
pandas
matplotlib.pyplot
math